LayoutTests/imported/w3c/web-platform-tests/fetch/api/request/request-keepalive.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request init: keep alive</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(function() {
          var requestBody = "body";
          assert_true(new Request("", {keepalive: true, body: requestBody, method: "POST"}).keepalive, "set keep alive to true");
          assert_false(new Request("", {keepalive: false, body: requestBody, method: "POST"}).keepalive, "set keep alive to false");
          assert_false(new Request("", {body: requestBody, method: "POST"}).keepalive, "keep alive default value");
      }, "Create Request with keep alive values");

      test(function() {
          var requestBody = "body";
          assert_true(new Request("", {keepalive: true, body: requestBody, method: "POST"}).clone().keepalive, "set keep alive to true");
          assert_false(new Request("", {keepalive: false, body: requestBody, method: "POST"}).clone().keepalive, "set keep alive to false");
          assert_false(new Request("", {body: requestBody, method: "POST"}).clone().keepalive, "keep alive default value");
      }, "Clone Request with keep alive values");

      test(function() {
          assert_throws(new TypeError(), () => {
            new Request("", {keepalive: true, body: new ReadableStream(), method: "POST"});
          });
      }, "Request with keep alive and ReadableStream body");
    </script>
  </body>
</html>
